chore: simplify testing configuration. restructurizing test utils
