+++ libraries/config/include/config/network.hpp
@@ -32,13 +32,13 @@ struct NetworkConfig {
 
   std::string json_file_name;
   std::string public_ip;
-  std::string listen_ip;
+  std::string listen_ip = "127.0.0.1";
   uint16_t listen_port = 0;
   std::vector<NodeConfig> boot_nodes;
-  uint16_t ideal_peer_count = 0;
-  uint16_t max_peer_count = 0;
-  uint16_t transaction_interval_ms = 0;
-  uint16_t sync_level_size = 0;
+  uint16_t ideal_peer_count = 10;
+  uint16_t max_peer_count = 50;
+  uint16_t transaction_interval_ms = 100;
+  uint16_t sync_level_size = 10;
   uint16_t performance_log_interval = 0;
   uint16_t num_threads = std::max(uint(1), uint(std::thread::hardware_concurrency() / 2));
   uint16_t packets_processing_threads = 14;
