+++ tests/py/config/wallet3.json
